tx20250120-1

+++ lib/ffi/sdr_funcs.go
@@ -1,10 +1,18 @@
 package ffi
 
 import (
+	"bufio"
 	"context"
 	"crypto/rand"
+	"encoding/gob"
 	"encoding/json"
+	"errors"
 	"fmt"
+	"github.com/filecoin-project/curio/lib/partialfile"
+	"github.com/filecoin-project/curio/txcar"
+	"github.com/filecoin-project/lotus/storage/sealer/fr32"
+	pool "github.com/libp2p/go-buffer-pool"
+	txcarlib "github.com/solopine/txcar/txcar"
 	"io"
 	"os"
 	"path/filepath"
@@ -29,6 +37,8 @@ import (
 
 	"github.com/filecoin-project/curio/lib/paths"
 	"github.com/filecoin-project/curio/lib/proofpaths"
+
+	carv2 "github.com/ipld/go-car/v2"
 )
 
 const C1CheckNumber = 3
@@ -269,7 +279,7 @@ func (sb *SealCalls) TreeRC(ctx context.Context, task *harmonytask.TaskID, secto
 				return cid.Undef, cid.Undef, xerrors.Errorf("truncating reflinked sealed file: %w", err)
 			}
 		} else {
-			log.Errorw("reflink treed -> sealed failed, falling back to slow copy, use single scratch btrfs or xfs filesystem", "error", err, "sector", sector, "cache", fspaths.Cache, "sealed", fspaths.Sealed)
+			log.Warnw("reflink treed -> sealed failed, falling back to slow copy, use single scratch btrfs or xfs filesystem", "error", err, "sector", sector, "cache", fspaths.Cache, "sealed", fspaths.Sealed)
 
 			// fallback to slow copy, copy ssize bytes from treed to sealed
 			dst, err := os.OpenFile(fspaths.Sealed, os.O_WRONLY|os.O_CREATE, 0644)
@@ -352,17 +362,19 @@ func (sb *SealCalls) GenerateSynthPoRep() {
 }
 
 func (sb *SealCalls) PoRepSnark(ctx context.Context, sn storiface.SectorRef, sealed, unsealed cid.Cid, ticket abi.SealRandomness, seed abi.InteractiveSealRandomness) ([]byte, error) {
+	log.Infow("----PoRepSnark.1", "sn", sn)
 	vproof, err := sb.sectors.storage.GeneratePoRepVanillaProof(ctx, sn, sealed, unsealed, ticket, seed)
 	if err != nil {
 		return nil, xerrors.Errorf("failed to generate vanilla proof: %w", err)
 	}
 
+	log.Infow("----PoRepSnark.2", "sn", sn)
 	ctx = ffiselect.WithLogCtx(ctx, "sector", sn.ID, "sealed", sealed, "unsealed", unsealed, "ticket", ticket, "seed", seed)
 	proof, err := ffiselect.FFISelect.SealCommitPhase2(ctx, vproof, sn.ID.Number, sn.ID.Miner)
 	if err != nil {
 		return nil, xerrors.Errorf("computing seal proof failed: %w", err)
 	}
-
+	log.Infow("----PoRepSnark.3", "sn", sn)
 	ok, err := ffi.VerifySeal(proof2.SealVerifyInfo{
 		SealProof:             sn.ProofType,
 		SectorID:              sn.ID,
@@ -525,7 +537,7 @@ func changePathType(path string, newType storiface.SectorFileType) (string, erro
 
 	return newPath, nil
 }
-func (sb *SealCalls) FinalizeSector(ctx context.Context, sector storiface.SectorRef, keepUnsealed bool) error {
+func (sb *SealCalls) FinalizeSector(ctx context.Context, sector storiface.SectorRef, keepUnsealed bool, txPiece *txcarlib.TxPiece) error {
 	sectorPaths, pathIDs, releaseSector, err := sb.sectors.AcquireSector(ctx, nil, sector, storiface.FTCache, storiface.FTNone, storiface.PathSealing)
 	if err != nil {
 		return xerrors.Errorf("acquiring sector paths: %w", err)
@@ -552,6 +564,13 @@ func (sb *SealCalls) FinalizeSector(ctx context.Context, sector storiface.Sector
 				log.Errorf("declare unsealed sector error: %+v", err)
 			}
 		}()
+		if txPiece != nil && txPiece.Version == txcarlib.TxCarV1 {
+			err := txcar.CreateFakeUnsealedFile(sectorPaths.Unsealed, *txPiece)
+			if err != nil {
+				return xerrors.Errorf("CreateFakeUnsealedFile - %s: %w", sectorPaths.Unsealed, err)
+			}
+			goto afterUnsealedMove
+		}
 
 		// tree-d contains exactly unsealed data in the prefix, so
 		// * we move it to a temp file
@@ -584,18 +603,36 @@ func (sb *SealCalls) FinalizeSector(ctx context.Context, sector storiface.Sector
 					}
 					return xerrors.Errorf("stat unsealed file: %w", err)
 				}
-				if st.Size() != int64(ssize) {
-					if tempUnsealedExists {
-						// unsealed file exists but is the wrong size, and temp unsealed file exists
-						// so we can just resume where the previous attempt left off with some cleanup
-
-						if err := os.Remove(sectorPaths.Unsealed); err != nil {
-							return xerrors.Errorf("removing unsealed file from last attempt: %w", err)
+				if txPiece != nil {
+					// tempUnsealed is not originate one
+					if !checkTxCarIndexFile(*txPiece, sectorPaths.Unsealed) {
+						log.Warnw("sectorPaths.Unsealed(%s) is not index file", sectorPaths.Unsealed)
+						// sectorPaths.Unsealed
+						if tempUnsealedExists {
+							// unsealed file exists but is not originate one, and temp unsealed file exists
+							// so we can just resume where the previous attempt left off with some cleanup
+							if err := os.Remove(sectorPaths.Unsealed); err != nil {
+								return xerrors.Errorf("removing unsealed file from last attempt: %w", err)
+							}
+							goto retryUnsealedMove
 						}
-
-						goto retryUnsealedMove
+						return xerrors.Errorf("sectorPaths.Unsealed(%s) is not txcarv2 index, and temp unsealed is missing: %w", sectorPaths.Unsealed, err)
 					}
-					return xerrors.Errorf("unsealed file is not the right size: %d != %d and temp unsealed is missing", st.Size(), ssize)
+					log.Warnw("sectorPaths.Unsealed(%s) is already index file", sectorPaths.Unsealed)
+				} else {
+					if st.Size() != int64(ssize) {
+						if tempUnsealedExists {
+							// unsealed file exists but is the wrong size, and temp unsealed file exists
+							// so we can just resume where the previous attempt left off with some cleanup
+
+							if err := os.Remove(sectorPaths.Unsealed); err != nil {
+								return xerrors.Errorf("removing unsealed file from last attempt: %w", err)
+							}
+
+							goto retryUnsealedMove
+						}
+						return xerrors.Errorf("unsealed file is not the right size: %d != %d and temp unsealed is missing", st.Size(), ssize)
+					}
 				}
 
 				// all good, just log that this edge case happened
@@ -619,9 +656,53 @@ func (sb *SealCalls) FinalizeSector(ctx context.Context, sector storiface.Sector
 			return xerrors.Errorf("truncating unsealed file to sector size: %w", err)
 		}
 
-		// move temp file to unsealed location
-		if err := paths.Move(tempUnsealed, sectorPaths.Unsealed); err != nil {
-			return xerrors.Errorf("move temp unsealed sector to final location (%s -> %s): %w", tempUnsealed, sectorPaths.Unsealed, err)
+		if txPiece != nil {
+
+			fsInfo, err := os.Stat(tempUnsealed)
+			if err != nil {
+				return xerrors.Errorf("os.Stat(%s): %w", tempUnsealed, err)
+			}
+			if fsInfo.Size() != int64(ssize) {
+				// tempUnsealed is not originate one
+				if !checkTxCarIndexFile(*txPiece, tempUnsealed) {
+					return xerrors.Errorf("tempUnsealed(%s) is not txcarv2 index, but size < ssize(%d): %w", tempUnsealed, int64(ssize), err)
+				} else {
+					log.Warnw("tempUnsealed(%s) is already txcarv2 index, just move to unsealed", tempUnsealed)
+					goto moveToUnsealed
+				}
+			}
+
+			{ // get index from unsealed
+				indexPath := tempUnsealed + ".index"
+
+				err = createTxCarIndexFileFromUnsealed(tempUnsealed, sector.ProofType, *txPiece, indexPath)
+				if err != nil {
+					return xerrors.Errorf("createTxCarV2IndexFileFromUnsealed: %w", err)
+				}
+
+				// remove tempUnsealed
+				err = os.Remove(tempUnsealed)
+				if err != nil {
+					return xerrors.Errorf("os.Remove(%s): %w", tempUnsealed, err)
+				}
+
+				err = os.Rename(indexPath, tempUnsealed)
+				if err != nil {
+					return xerrors.Errorf("os.Rename(%s -> %s): %w", indexPath, tempUnsealed, err)
+				}
+			}
+
+		moveToUnsealed:
+			// move file to unsealed location
+			if err := paths.Move(tempUnsealed, sectorPaths.Unsealed); err != nil {
+				return xerrors.Errorf("move temp unsealed index sector to final location (%s -> %s): %w", tempUnsealed, sectorPaths.Unsealed, err)
+			}
+
+		} else {
+			// move temp file to unsealed location
+			if err := paths.Move(tempUnsealed, sectorPaths.Unsealed); err != nil {
+				return xerrors.Errorf("move temp unsealed sector to final location (%s -> %s): %w", tempUnsealed, sectorPaths.Unsealed, err)
+			}
 		}
 	}
 
@@ -789,4 +870,114 @@ func (sb *SealCalls) SyntheticProofs(ctx context.Context, task *harmonytask.Task
 	}
 
 	return nil
+}
+
+var ReadBuf = 128 * (127 * 8)
+
+func createTxCarIndexFileFromUnsealed(unsealedPath string, proofType abi.RegisteredSealProof, txPiece txcarlib.TxPiece, indexPath string) error {
+	var br io.Reader
+	{
+		ssize, err := proofType.SectorSize()
+		if err != nil {
+			return err
+		}
+
+		pieceSize := abi.PaddedPieceSize(ssize)
+
+		// open the unsealed sector file for the given sector size located at the given path.
+		pf, err := partialfile.OpenPartialFile(pieceSize, unsealedPath)
+		if err != nil {
+			return xerrors.Errorf("opening partial file: %w", err)
+		}
+
+		defer func() {
+			err := pf.Close()
+			if err != nil {
+				log.Errorw("pf.Close()", "unsealedPath", unsealedPath)
+			}
+		}()
+
+		r, err := pf.Reader(0, pieceSize)
+		if err != nil {
+			return err
+		}
+
+		buf := pool.Get(fr32.BufSize(pieceSize))
+
+		upr0, err := fr32.NewUnpadReaderBuf(r, pieceSize, buf)
+		if err != nil {
+			return xerrors.Errorf("creating unpadded reader: %w", err)
+		}
+
+		upr := io.LimitReader(upr0, int64(pieceSize.Unpadded()))
+
+		bir := bufio.NewReaderSize(upr, 127)
+		br = bufio.NewReaderSize(bir, ReadBuf)
+	}
+
+	// generate recs
+	txRecs := make([]txcarlib.TxBlockRecord, 0, 10000)
+	opts := []carv2.Option{carv2.ZeroLengthSectionAsEOF(true)}
+	blockReader, err := carv2.NewBlockReader(br, opts...)
+	if err != nil {
+		return fmt.Errorf("getting block reader over piece: %w", err)
+	}
+
+	blockMetadata, err := blockReader.SkipNext()
+	for err == nil {
+		txRecs = append(txRecs, txcarlib.TxBlockRecord{
+			Cid:    blockMetadata.Cid,
+			Offset: blockMetadata.SourceOffset,
+			Size:   blockMetadata.Size,
+		})
+
+		blockMetadata, err = blockReader.SkipNext()
+	}
+	if !errors.Is(err, io.EOF) {
+		return fmt.Errorf("generating index for piece: %w", err)
+	}
+
+	// write file
+	indexFile, err := os.Create(indexPath)
+	if err != nil {
+		return err
+	}
+
+	enc := gob.NewEncoder(indexFile)
+	// 1. write txpiece
+	err = enc.Encode(txPiece)
+	if err != nil {
+		return err
+	}
+
+	// 2. write txRecs
+	err = enc.Encode(txRecs)
+	if err != nil {
+		return err
+	}
+
+	return indexFile.Close()
+}
+
+func checkTxCarIndexFile(txPiece txcarlib.TxPiece, indexPath string) bool {
+	reader, err := os.Open(indexPath)
+	if err != nil {
+		return false
+	}
+
+	dec := gob.NewDecoder(reader)
+
+	// 1. read txpiece
+	var txPieceRec txcarlib.TxPiece
+	if err := dec.Decode(&txPieceRec); err != nil {
+		return false
+	}
+	if txPiece.PieceCid != txPieceRec.PieceCid {
+		return false
+	}
+	if txPiece.Version != txPieceRec.Version {
+		return false
+	}
+
+	return true
 }
